+++ R/cross_validate_list.R
@@ -397,8 +397,8 @@ cross_validate_list <- function(data,
       # Evaluate the predictions
       prediction_evaluation <- internal_evaluate_predictions(
         data = current_predictions,
-        predictions_col = "prediction",
-        targets_col = "target",
+        prediction_col = "prediction",
+        target_col = "target",
         model_was_null_col = "model_was_null",
         type = family,
         fold_info_cols = fold_info_cols,