+++ R/plot_probabilities.R
@@ -70,7 +70,7 @@
 #'
 #'  As created with the various validation functions in \code{cvms}, like
 #'  \code{\link[cvms:cross_validate_fn]{cross_validate_fn()}}.
-#' @param targets_col Name of column with target levels.
+#' @param target_col Name of column with target levels.
 #' @param probability_cols Name of columns with predicted probabilities.
 #'
 #'  For \strong{binary} classification, this should be \strong{one column} with the probability of the